Complete login and sign up implementation (#13)
* [fix] remove duplicated log

* [feat] add password field

* [feat] add calling to create user

* [fix] merge conflict

* [refactor] change long to int

* [feat] add create user

* [feat] remove default log

* [style] format code

* [feat] create login request and response DTOs, update UserAPIService

* [feat] add androidx.security:security-crypto dependency

* [feat] implement SharedPrefUtil for encrypted shared preferences management

* [feat] add ErrorInfo and GenericApiResponse classes for API error handling

* [feat] implement login functionality with API integration and error handling

* [build] add .idea in gitignore

* [refactor] comment login logic out (temporary)

+++ app/src/main/java/com/example/travelexpertmobileapplication/network/api/AgentAPIService.java
@@ -17,7 +17,7 @@ public interface AgentAPIService {
     @Multipart
     @POST("/agents/{id}/upload")
     Call<ResponseBody> uploadAgentPhoto(
-            @Path("id") Long agentId,
+            @Path("id") int agentId,
             @Part MultipartBody.Part image
     );
 }
